analysis_crew_js/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.30
botocore==1.35.30
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
clarifai==10.8.0
clarifai-grpc==10.8.7
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.28.8
dataclasses-json==0.6.7
Deprecated==1.2.14
dill==0.3.9
distro==1.9.0
docker==7.1.0
docstring-parser==0.15
durationpy==0.8
embedchain==0.1.113
exceptiongroup==1.2.2
fastapi==0.115.0
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-aiplatform==1.68.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.64.3
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inquirerpy==0.3.4
instructor==0.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==31.0.0
langchain==0.1.20
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.57
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.129
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==5.0.1
mock==4.0.3
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.50.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
pathos==0.3.3
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
posthog==3.6.6
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pulsar-client==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-rapidjson==1.20
pytz==2024.2
pywin32==306
PyYAML==6.0.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
sagemaker==2.232.1
sagemaker-core==1.0.9
schema==0.7.5
scikit-learn==1.4.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tomli==2.0.1
tqdm==4.66.5
tritonclient==2.50.0
typer==0.9.4
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
